+++ tomcat/src/main/java/org/apache/coyote/http11/request/Request.java
@@ -1,7 +1,5 @@
 package org.apache.coyote.http11.request;
 
-import org.apache.coyote.http11.request.body.RequestBody;
-import org.apache.coyote.http11.request.header.RequestHeader;
 import org.apache.coyote.http11.request.line.RequestLine;
 
 public class Request {
