+++ tomcat/src/main/java/org/apache/coyote/http11/request/RequestBody.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request.body;
+package org.apache.coyote.http11.request;
 
 import java.util.Arrays;
 import java.util.Map;
