+++ tomcat/src/main/java/org/apache/coyote/http11/request/line/RequestLine.java
@@ -5,6 +5,10 @@ import java.util.List;
 
 public class RequestLine {
 
+    private static final int HTTP_METHOD_INDEX = 0;
+    private static final int PATH_INDEX = 1;
+    private static final int PROTOCOL_INDEX = 2;
+
     private final HttpMethod method;
     private final Path path;
     private final Protocol protocol;
@@ -17,9 +21,9 @@ public class RequestLine {
 
     public static RequestLine from(String requestLine) {
         List<String> requestLineElements = Arrays.asList(requestLine.split(" "));
-        HttpMethod method = HttpMethod.from(requestLineElements.get(0));
-        Path path = Path.from(requestLineElements.get(1));
-        Protocol protocol = Protocol.from(requestLineElements.get(2));
+        HttpMethod method = HttpMethod.from(requestLineElements.get(HTTP_METHOD_INDEX));
+        Path path = Path.from(requestLineElements.get(PATH_INDEX));
+        Protocol protocol = Protocol.from(requestLineElements.get(PROTOCOL_INDEX));
         return new RequestLine(method, path, protocol);
     }
