+++ tomcat/src/test/java/org/apache/coyote/http11/auth/AuthServiceTest.java
@@ -3,8 +3,8 @@ package org.apache.coyote.http11.auth;
 import java.util.List;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.request.body.RequestBody;
-import org.apache.coyote.http11.request.header.RequestHeader;
+import org.apache.coyote.http11.request.RequestBody;
+import org.apache.coyote.http11.request.RequestHeader;
 import org.apache.coyote.http11.request.line.HttpMethod;
 import org.apache.coyote.http11.request.line.RequestLine;
 import org.apache.coyote.http11.response.ResponseEntity;
