+++ tomcat/src/test/java/org/apache/coyote/http11/request/handler/RequestHandlerTest.java
@@ -3,12 +3,11 @@ package org.apache.coyote.http11.request.handler;
 import java.util.List;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.auth.Cookie;
-import org.apache.coyote.http11.auth.Session;
 import org.apache.coyote.http11.auth.SessionRepository;
 import org.apache.coyote.http11.request.Request;
-import org.apache.coyote.http11.request.body.RequestBody;
-import org.apache.coyote.http11.request.header.RequestHeader;
+import org.apache.coyote.http11.request.RequestBody;
+import org.apache.coyote.http11.request.RequestHandler;
+import org.apache.coyote.http11.request.RequestHeader;
 import org.apache.coyote.http11.request.line.HttpMethod;
 import org.apache.coyote.http11.request.line.RequestLine;
 import org.apache.coyote.http11.response.ResponseEntity;
